fix: narrow unrecognized provider refusal fallback
