+++ pkg/services/workers/internal/services/runners/internal/services/agent/internal/service/runner_unrecognized_failure_test.go
@@ -44,6 +44,9 @@ func TestExecuteUnrecognizedProviderFailureIsTerminalAndProviderNeutral(t *testi
 	fake := &failingProvidersFake{failure: providers.ExecuteFailure{
 		Kind:    providers.ExecuteFailureKindUnknown,
 		Message: providerDetail,
+		Diagnostics: &providers.ExecuteDiagnostics{Metadata: map[string]string{
+			providers.ExecuteDiagnosticMetadataUnrecognizedProviderRefusal: "true",
+		}},
 	}}
 	runner, err := New(fake, noopPublisher)
 	if err != nil {
