+++ pkg/services/workers/internal/services/workstations/invocation/executor_root_test.go
@@ -135,6 +135,30 @@ func TestProviderExecutorExecutePreservesProviderMarkedSafeFailureMessage(t *tes
 	}
 }
 
+func TestProviderExecutorExecuteRoutesMarkedUnknownRefusalAsPermanent(t *testing.T) {
+	providerErr := providers.ExecuteFailure{
+		Kind:    providers.ExecuteFailureKindUnknown,
+		Message: "future provider refusal",
+		Diagnostics: &providers.ExecuteDiagnostics{Metadata: map[string]string{
+			providers.ExecuteDiagnosticMetadataUnrecognizedProviderRefusal: "true",
+		}},
+	}
+
+	result, err := workerinvocation.NewProviderExecutor(&executionTestProvider{err: providerErr}).Execute(
+		context.Background(),
+		workerexecution.InvocationInput{},
+	)
+	if err == nil {
+		t.Fatal("Execute() error = nil, want marked provider refusal")
+	}
+	if result.FailureDetail == nil || result.FailureDetail.Reason != workerexecution.WorkFailureTypePermanentBadRequest {
+		t.Fatalf("failure detail = %#v, want permanent bad request", result.FailureDetail)
+	}
+	if result.FailureDecision == nil || !result.FailureDecision.Terminal || result.FailureDecision.Retryable {
+		t.Fatalf("failure decision = %#v, want terminal non-retryable refusal", result.FailureDecision)
+	}
+}
+
 func TestProviderExecutorExecutePropagatesCancellationWithoutRetry(t *testing.T) {
 	provider := newBlockingExecutionTestProvider()
 	ctx, cancel := context.WithCancel(context.Background())
@@ -592,6 +616,8 @@ func assertProviderFailureKinds(t *testing.T) {
 		{providers.ExecuteFailureKindThrottled, workerexecution.WorkFailureTypeThrottled},
 		{providers.ExecuteFailureKindMisconfigured, workerexecution.WorkFailureTypeMisconfigured},
 		{providers.ExecuteFailureKindDependency, workerexecution.WorkFailureTypeUnknown},
+		{providers.ExecuteFailureKindUnknown, workerexecution.WorkFailureTypeUnknown},
+		{providers.ExecuteFailureKindSessionNotFound, workerexecution.WorkFailureTypeUnknown},
 	}
 	for _, testCase := range failureKinds {
 		t.Run(string(testCase.kind), func(t *testing.T) {
